230730 로그인, 회원가입, 유효성 검사 구현, 시큐리티 적용중

+++ src/main/java/com/example/project_collatool/dto/BoardDto.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BoardDto {
-    private Long bId;
-    private Long bProjectId;
-    private Long bUserId;
+    private Integer bId;
+    private Integer bProjectId;
+    private Integer bUserId;
     private String bTitle;
     private String bContent;
     private String bFname;
